Refactor : 스케줄러 기능 수정
