+++ src/main/java/com/project/team11_tabling/domain/shop/entity/ShopSeats.java
@@ -35,4 +35,7 @@ public class ShopSeats {
         .build();
   }
 
+  public void addAvailableSeat() {
+    this.availableSeats ++;
+  }
 }
