+++ src/main/java/com/project/team11_tabling/global/config/SchedulerConfig.java
@@ -1,4 +1,4 @@
-package com.project.team11_tabling.global.scheduler;
+package com.project.team11_tabling.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.SchedulingConfigurer;
@@ -6,7 +6,7 @@ import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
 import org.springframework.scheduling.config.ScheduledTaskRegistrar;
 
 @Configuration
-public class TaskSchedulerConfiguration implements SchedulingConfigurer {
+public class SchedulerConfig implements SchedulingConfigurer {
   @Override
   public void configureTasks(ScheduledTaskRegistrar taskRegistrar) {
     // Thread Pool 설정
